LPS-167508 Move ObjectEntryEntityModel back to object-rest-impl as is an implementation and for consistency with the rest of EntityModels
