+++ modules/apps/object/object-rest-api/bnd.bnd
@@ -6,7 +6,6 @@ Export-Package:\
 	com.liferay.object.rest.dto.v1_0.util,\
 	com.liferay.object.rest.manager.exception,\
 	com.liferay.object.rest.manager.v1_0,\
-	com.liferay.object.rest.odata.entity.v1_0,\
 	com.liferay.object.rest.openapi.v1_0,\
 	com.liferay.object.rest.petra.sql.dsl.expression,\
 	com.liferay.object.rest.resource.v1_0
