+++ modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/odata/entity/v1_0/ObjectEntryEntityModel.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.object.rest.odata.entity.v1_0;
+package com.liferay.object.rest.internal.odata.entity.v1_0;
 
 import com.liferay.object.constants.ObjectFieldConstants;
 import com.liferay.object.constants.ObjectFieldSettingConstants;
